refactor(commands): extract python and project command modules (Issue #271)
Move handle_python_command (and helpers) into commands/python.rs, and
init_project/run_drift/sanitize_project_name into commands/project.rs,
following the same move-only extraction discipline used for
commands/tooling.rs. Also drops unused imports left behind in
commands/mod.rs after the extraction.

No behavior changes: JSON envelope and MCP wire format are unchanged.
