+++ src/commands/mod.rs
@@ -1,22 +1,16 @@
 use crate::build::{BuildBackend, BuildCache};
 use crate::cli::{
-    Cli, Commands, DriftArgs, InitArgs, InitTemplate, McpCommands, OutputFormat, ProgressMode,
-    PythonCommands, SchemaArgs, SchemaCommands, SelfCommands, TelemetryCommands,
+    Cli, Commands, McpCommands, OutputFormat, ProgressMode, PythonCommands, SchemaArgs,
+    SchemaCommands, SelfCommands, TelemetryCommands,
 };
 use crate::env::find_python_env;
 use crate::progress::{ProgressConfig, ProgressDriver};
 use crate::project::Project;
-use crate::release_manifest::{ReleaseManifest, current_release_target};
 use crate::sandbox;
 use crate::sbom;
-use crate::schema::{
-    Diagnostic, DiagnosticLevel, Event, EventCollector, EventType, JsonEnvelope, Status,
-};
-use crate::self_update::apply_update_for_asset;
+use crate::schema::{Diagnostic, Event, EventCollector, EventType, JsonEnvelope, Status};
 use color_eyre::eyre::{Result, eyre};
-use dialoguer::{Input, theme::ColorfulTheme};
 use serde_json::{Value, json};
-use std::cmp::Ordering;
 use std::fs;
 use std::io::IsTerminal;
 use std::path::{Path, PathBuf};
@@ -27,6 +21,8 @@ mod install;
 use install::{AddOutcome, LockOutcome, RemoveOutcome};
 pub(crate) use install::{InstallOutcome, RunOutcome, install};
 mod maintenance;
+mod project;
+mod python;
 mod run;
 use run::{SandboxInfo, get_python_version, script_lock_path};
 pub(crate) use run::{find_python_interpreter, run_script};
@@ -511,9 +507,9 @@ pub async fn execute(cli: Cli) -> Result<()> {
         }
         Commands::X(args) => {
             collector.event(EventType::EnvCreate);
-            let result = execute_tool(args, &mut collector);
+            let result = tooling::execute_tool(args, &mut collector);
             match result {
-                Ok(XOutcome {
+                Ok(tooling::XOutcome {
                     summary,
                     package,
                     version,
@@ -682,7 +678,7 @@ pub async fn execute(cli: Cli) -> Result<()> {
         },
         Commands::SelfCmd(cmd) => match cmd {
             SelfCommands::Update(args) => {
-                let detail = run_self_update(args, &mut collector);
+                let detail = tooling::run_self_update(args, &mut collector);
                 ("self update".to_string(), detail)
             }
         },
@@ -710,7 +706,7 @@ pub async fn execute(cli: Cli) -> Result<()> {
             }
         }
         Commands::Python(cmd) => {
-            match handle_python_command(cmd, &mut collector) {
+            match python::handle_python_command(cmd, &mut collector) {
                 Ok((subcmd, detail)) => (format!("python {}", subcmd), detail),
                 Err(e) => {
                     // Determine subcommand name for error reporting
@@ -882,7 +878,7 @@ pub async fn execute(cli: Cli) -> Result<()> {
         }
         Commands::Init(args) => {
             let pre_error_count = collector.error_diagnostic_count();
-            let result = init_project(args, &mut collector);
+            let result = project::init_project(args, &mut collector);
             match result {
                 Ok(detail) => ("init".to_string(), detail),
                 Err(e) => {
@@ -962,7 +958,7 @@ pub async fn execute(cli: Cli) -> Result<()> {
             }
         }
         Commands::Drift(args) => {
-            let result = run_drift(args, &mut collector);
+            let result = project::run_drift(args, &mut collector);
             match result {
                 Ok(detail) => ("drift".to_string(), detail),
                 Err(e) => {
@@ -1481,981 +1477,5 @@ fn emit_rejected_allow_env_diagnostics(collector: &mut EventCollector, rejected_
             ))
             .with_code("W_SANDBOX_ALLOW_ENV_REJECTED"),
         );
-    }
-}
-
-// ---------------------------------------------------------------------------
-// pybun python
-// ---------------------------------------------------------------------------
-
-use crate::cache::Cache;
-use crate::runtime::{RuntimeManager, supported_versions};
-
-fn handle_python_command(
-    cmd: &PythonCommands,
-    collector: &mut EventCollector,
-) -> Result<(String, RenderDetail)> {
-    match cmd {
-        PythonCommands::List(args) => {
-            collector.event(EventType::PythonListStart);
-            let result = python_list(args);
-            collector.event(EventType::PythonListComplete);
-            result
-        }
-        PythonCommands::Install(args) => {
-            collector.event(EventType::PythonInstallStart);
-            let result = python_install(args);
-            collector.event(EventType::PythonInstallComplete);
-            result
-        }
-        PythonCommands::Remove(args) => {
-            collector.event(EventType::PythonRemoveStart);
-            let result = python_remove(args);
-            collector.event(EventType::PythonRemoveComplete);
-            result
-        }
-        PythonCommands::Which(args) => python_which(args),
-    }
-}
-
-fn python_list(args: &crate::cli::PythonListArgs) -> Result<(String, RenderDetail)> {
-    let cache = Cache::new().map_err(|e| eyre!("failed to initialize cache: {}", e))?;
-    let manager = RuntimeManager::new(cache);
-
-    let installed = manager.list_installed()?;
-    let available = supported_versions();
-
-    let mut text_output = String::new();
-
-    if args.all {
-        text_output.push_str("Available Python versions:\n");
-        for v in &available {
-            let status = if installed.iter().any(|i| i == &v.version) {
-                " (installed)"
-            } else {
-                ""
-            };
-            text_output.push_str(&format!("  {}{}\n", v.version, status));
-        }
-    } else {
-        text_output.push_str("Installed Python versions:\n");
-        if installed.is_empty() {
-            text_output.push_str("  (none)\n");
-            text_output
-                .push_str("\nUse 'pybun python install <VERSION>' to install a Python version.");
-        } else {
-            for v in &installed {
-                text_output.push_str(&format!("  {}\n", v));
-            }
-        }
-    }
-
-    let json = json!({
-        "installed": installed,
-        "available": available.iter().map(|v| &v.version).collect::<Vec<_>>(),
-    });
-
-    Ok((
-        "list".to_string(),
-        RenderDetail::with_json(text_output.trim(), json),
-    ))
-}
-
-fn python_install(args: &crate::cli::PythonInstallArgs) -> Result<(String, RenderDetail)> {
-    let cache = Cache::new().map_err(|e| eyre!("failed to initialize cache: {}", e))?;
-    let manager = RuntimeManager::new(cache);
-
-    // Check if already installed
-    if manager.is_installed(&args.version) {
-        let path = manager.python_binary(&args.version);
-        let summary = format!(
-            "Python {} is already installed at {}",
-            args.version,
-            path.display()
-        );
-        let json = json!({
-            "version": args.version,
-            "path": path.display().to_string(),
-            "status": "already_installed",
-        });
-        return Ok((
-            "install".to_string(),
-            RenderDetail::with_json(summary, json),
-        ));
-    }
-
-    // Install
-    let python_path = manager.ensure_version(&args.version)?;
-
-    let summary = format!(
-        "Installed Python {} at {}",
-        args.version,
-        python_path.display()
-    );
-    let json = json!({
-        "version": args.version,
-        "path": python_path.display().to_string(),
-        "status": "installed",
-    });
-
-    Ok((
-        "install".to_string(),
-        RenderDetail::with_json(summary, json),
-    ))
-}
-
-fn python_remove(args: &crate::cli::PythonRemoveArgs) -> Result<(String, RenderDetail)> {
-    let cache = Cache::new().map_err(|e| eyre!("failed to initialize cache: {}", e))?;
-    let manager = RuntimeManager::new(cache);
-
-    manager.remove_version(&args.version)?;
-
-    let summary = format!("Removed Python {}", args.version);
-    let json = json!({
-        "version": args.version,
-        "status": "removed",
-    });
-
-    Ok(("remove".to_string(), RenderDetail::with_json(summary, json)))
-}
-
-// ---------------------------------------------------------------------------
-// pybun x (execute tool ad-hoc)
-// ---------------------------------------------------------------------------
-
-#[derive(Debug)]
-struct XOutcome {
-    summary: String,
-    package: String,
-    version: Option<String>,
-    passthrough: Vec<String>,
-    temp_env: String,
-    python_version: String,
-    exit_code: i32,
-    cleanup: bool,
-}
-
-fn execute_tool(args: &crate::cli::ToolArgs, _collector: &mut EventCollector) -> Result<XOutcome> {
-    let package_spec = args
-        .package
-        .as_ref()
-        .ok_or_else(|| eyre!("package name is required"))?;
-
-    // Parse package name and version
-    let (package_name, version) = parse_package_spec(package_spec);
-
-    // Check for dry-run mode (for testing)
-    let dry_run = std::env::var("PYBUN_X_DRY_RUN").is_ok();
-
-    // Find Python interpreter
-    let working_dir = std::env::current_dir()?;
-    let env = find_python_env(&working_dir)?;
-    let python_path = env.python_path.to_string_lossy().to_string();
-    let python_version = env.version.clone().unwrap_or_else(|| "unknown".to_string());
-
-    // Create temporary environment
-    let temp_dir =
-        tempfile::tempdir().map_err(|e| eyre!("failed to create temp directory: {}", e))?;
-    let temp_env_path = temp_dir.path().to_string_lossy().to_string();
-
-    if dry_run {
-        // In dry-run mode, just return the planned actions. Tests can set
-        // PYBUN_X_DRY_RUN_EXIT_CODE to simulate a tool that exits non-zero
-        // without needing network access / a real pip install.
-        let exit_code = std::env::var("PYBUN_X_DRY_RUN_EXIT_CODE")
-            .ok()
-            .and_then(|v| v.parse().ok())
-            .unwrap_or(0);
-        return Ok(XOutcome {
-            summary: format!("would execute {} (dry-run)", package_name),
-            package: package_name,
-            version,
-            passthrough: args.passthrough.clone(),
-            temp_env: temp_env_path,
-            python_version,
-            exit_code,
-            cleanup: true,
-        });
-    }
-
-    // Create virtual environment in temp directory
-    let venv_path = temp_dir.path().join("venv");
-    eprintln!(
-        "info: creating temporary environment at {}",
-        venv_path.display()
-    );
-
-    let uv_available = crate::env::find_uv_executable().is_some();
-
-    let mut venv_cmd = ProcessCommand::new(&python_path);
-    venv_cmd.args(["-m", "venv"]);
-    if uv_available {
-        venv_cmd.arg("--without-pip");
-    }
-    venv_cmd.arg(&venv_path);
-    let venv_status = venv_cmd
-        .status()
-        .map_err(|e| eyre!("failed to create virtual environment: {}", e))?;
-
-    if !venv_status.success() {
-        return Err(eyre!("failed to create virtual environment"));
-    }
-
-    // Get pip path in venv
-    let pip_path = if cfg!(windows) {
-        venv_path.join("Scripts").join("pip.exe")
-    } else {
-        venv_path.join("bin").join("pip")
-    };
-
-    // Get python path in venv
-    let venv_python = if cfg!(windows) {
-        venv_path.join("Scripts").join("python.exe")
-    } else {
-        venv_path.join("bin").join("python")
-    };
-
-    // Install the package using uv if available, otherwise pip
-    eprintln!("info: installing {}...", package_spec);
-    let install_status = if let Some(uv_path) = crate::env::find_uv_executable() {
-        ProcessCommand::new(uv_path)
-            .args(["pip", "install", "--quiet", "--python"])
-            .arg(&venv_path)
-            .arg(package_spec)
-            .status()
-            .map_err(|e| eyre!("failed to install package with uv: {}", e))?
-    } else {
-        ProcessCommand::new(&pip_path)
-            .args(["install", "--quiet", package_spec])
-            .status()
-            .map_err(|e| eyre!("failed to install package: {}", e))?
-    };
-
-    if !install_status.success() {
-        return Err(eyre!("failed to install package {}", package_spec));
-    }
-
-    // Find and execute the entry point
-    // Most packages have a console script with the same name as the package
-    let entry_point = if cfg!(windows) {
-        venv_path
-            .join("Scripts")
-            .join(format!("{}.exe", package_name))
-    } else {
-        venv_path.join("bin").join(&package_name)
-    };
-
-    let exit_code = if entry_point.exists() {
-        // Execute the console script directly
-        eprintln!("info: executing {}...", entry_point.display());
-        let mut cmd = ProcessCommand::new(&entry_point);
-        for arg in &args.passthrough {
-            cmd.arg(arg);
-        }
-        let status = cmd
-            .status()
-            .map_err(|e| eyre!("failed to execute {}: {}", package_name, e))?;
-        status.code().unwrap_or(-1)
-    } else {
-        // Fallback: try to run as a module
-        eprintln!("info: executing python -m {}...", package_name);
-        let mut cmd = ProcessCommand::new(&venv_python);
-        cmd.args(["-m", &package_name]);
-        for arg in &args.passthrough {
-            cmd.arg(arg);
-        }
-        let status = cmd
-            .status()
-            .map_err(|e| eyre!("failed to execute module {}: {}", package_name, e))?;
-        status.code().unwrap_or(-1)
-    };
-
-    // Cleanup is automatic when temp_dir is dropped
-    let summary = if exit_code == 0 {
-        format!("executed {} successfully", package_name)
-    } else {
-        format!("{} exited with code {}", package_name, exit_code)
-    };
-
-    Ok(XOutcome {
-        summary,
-        package: package_name,
-        version,
-        passthrough: args.passthrough.clone(),
-        temp_env: temp_env_path,
-        python_version,
-        exit_code,
-        cleanup: true,
-    })
-}
-
-/// Parse a package specification like "cowsay==6.1" into (name, version)
-fn parse_package_spec(spec: &str) -> (String, Option<String>) {
-    // Handle various specifier formats
-    for sep in ["==", ">=", "<=", "!=", "~=", ">", "<"] {
-        if let Some(idx) = spec.find(sep) {
-            let name = spec[..idx].to_string();
-            let version = spec[idx + sep.len()..].to_string();
-            return (name, Some(version));
-        }
-    }
-    (spec.to_string(), None)
-}
-
-// ---------------------------------------------------------------------------
-// pybun self update
-// ---------------------------------------------------------------------------
-
-fn run_self_update(
-    args: &crate::cli::SelfUpdateArgs,
-    collector: &mut EventCollector,
-) -> RenderDetail {
-    let current_version = env!("CARGO_PKG_VERSION");
-    let channel = &args.channel;
-
-    collector.info(format!("Checking for updates on {} channel", channel));
-
-    let manifest_source_env = std::env::var("PYBUN_SELF_UPDATE_MANIFEST").ok();
-    let default_manifest_url = default_manifest_url(channel);
-    let manifest_source = manifest_source_env
-        .clone()
-        .unwrap_or_else(|| default_manifest_url.clone());
-    let should_fetch_manifest = !args.dry_run
-        || manifest_source_env.is_some()
-        || std::env::var("PYBUN_SELF_UPDATE_FETCH").is_ok();
-    let manifest_result = if should_fetch_manifest {
-        Some(ReleaseManifest::load(&manifest_source))
-    } else {
-        None
-    };
-
-    let mut latest_version = current_version.to_string();
-    let mut update_available = false;
-    let mut release_url = release_url_for_version(current_version);
-    let target = current_release_target();
-    let mut selected_asset = None;
-    let mut manifest_detail = None;
-    let mut manifest_error = None;
-    let mut update_applied = false;
-    let mut rollback_performed = false;
-    let mut install_path = None;
-    let mut update_error = None;
-
-    match manifest_result {
-        Some(Ok(manifest)) => {
-            latest_version = manifest.version.clone();
-            update_available = manifest
-                .compare_version(current_version)
-                .map(|ordering| ordering == Ordering::Greater)
-                .unwrap_or(false);
-            release_url = manifest
-                .release_url
-                .clone()
-                .unwrap_or_else(|| release_url_for_version(&manifest.version));
-
-            selected_asset = target
-                .as_deref()
-                .and_then(|target| manifest.select_asset(target))
-                .cloned();
-            let asset_json = selected_asset
-                .as_ref()
-                .map(|asset| serde_json::to_value(asset).unwrap_or_else(|_| json!({})))
-                .unwrap_or(Value::Null);
-
-            manifest_detail = Some(json!({
-                "version": manifest.version,
-                "channel": manifest.channel,
-                "published_at": manifest.published_at,
-                "release_url": manifest.release_url,
-                "release_notes": manifest.release_notes,
-                "source": manifest_source,
-                "target": target,
-                "asset": asset_json,
-                "assets": manifest.assets.len(),
-                "sbom": manifest.sbom,
-                "provenance": manifest.provenance,
-            }));
-        }
-        Some(Err(error)) => {
-            manifest_error = Some(error.to_string());
-        }
-        None => {}
-    }
-
-    if !args.dry_run {
-        if let Some(error) = manifest_error.as_deref() {
-            let message = format!("failed to load release manifest: {error}");
-            collector.error_with_code(
-                "E_SELF_UPDATE_MANIFEST",
-                message.clone(),
-                "Check network connectivity and the release manifest URL (--channel or PYBUN_SELF_UPDATE_MANIFEST_URL), then retry `pybun self update`.",
-            );
-            update_error = Some(message);
-        } else if update_available {
-            let Some(asset) = selected_asset else {
-                let target_text = target
-                    .as_deref()
-                    .map(|value| value.to_string())
-                    .unwrap_or_else(|| "unknown".to_string());
-                let message = format!("no release asset found for target {target_text}");
-                collector.error_with_code(
-                    "E_SELF_UPDATE_NO_ASSET",
-                    message.clone(),
-                    "Pass --target explicitly to select a supported platform/arch, or check that a release asset exists for your platform.",
-                );
-                update_error = Some(message);
-                let summary = "Update failed: no release asset found".to_string();
-                let json_detail = json!({
-                    "current_version": current_version,
-                    "latest_version": latest_version,
-                    "channel": channel,
-                    "update_available": update_available,
-                    "release_url": release_url,
-                    "dry_run": args.dry_run,
-                    "target": target,
-                    "manifest": manifest_detail,
-                    "manifest_error": manifest_error,
-                    "manifest_source": manifest_source_env.or(Some(default_manifest_url)),
-                    "update_applied": false,
-                    "rollback_performed": false,
-                    "install_path": Value::Null,
-                    "error": update_error,
-                });
-                return RenderDetail::error(summary, json_detail);
-            };
-            let install_override = std::env::var("PYBUN_SELF_UPDATE_BIN")
-                .ok()
-                .map(PathBuf::from);
-            let fail_swap_for_test = std::env::var("PYBUN_SELF_UPDATE_TEST_FAIL_SWAP").is_ok();
-            let target_name = target
-                .as_deref()
-                .unwrap_or(asset.target.as_str())
-                .to_string();
-
-            match apply_update_for_asset(&asset, &target_name, install_override, fail_swap_for_test)
-            {
-                Ok(outcome) => {
-                    update_applied = true;
-                    rollback_performed = outcome.rollback_performed;
-                    install_path = Some(outcome.install_path.display().to_string());
-                    collector.info(format!(
-                        "Updated binary at {}",
-                        outcome.install_path.display()
-                    ));
-                }
-                Err(error) => {
-                    rollback_performed = error.rollback_performed;
-                    update_error = Some(error.to_string());
-                    collector.error_with_code(
-                        "E_SELF_UPDATE_APPLY_FAILED",
-                        error.to_string(),
-                        "Check write permissions to the install path and retry `pybun self update`. If a backup exists, rollback may have already been performed.",
-                    );
-                }
-            }
-        }
-    }
-
-    let summary = if args.dry_run {
-        if update_available {
-            format!(
-                "Update available: {} -> {} (dry-run, no changes made)",
-                current_version, latest_version
-            )
-        } else if let Some(error) = manifest_error.as_deref() {
-            format!("Update check failed: {} (dry-run)", error)
-        } else {
-            format!(
-                "Already up to date: {} (channel: {})",
-                current_version, channel
-            )
-        }
-    } else if let Some(error) = update_error.as_deref() {
-        format!("Update failed: {error}")
-    } else if update_available && update_applied {
-        format!("Updated: {} -> {}", current_version, latest_version)
-    } else if update_available {
-        format!("Update failed: {} -> {}", current_version, latest_version)
-    } else {
-        format!(
-            "Already up to date: {} (channel: {})",
-            current_version, channel
-        )
-    };
-
-    let json_detail = json!({
-        "current_version": current_version,
-        "latest_version": latest_version,
-        "channel": channel,
-        "update_available": update_available,
-        "release_url": release_url,
-        "dry_run": args.dry_run,
-        "target": target,
-        "manifest": manifest_detail,
-        "manifest_error": manifest_error,
-        "manifest_source": manifest_source_env.or(Some(default_manifest_url)),
-        "update_applied": update_applied,
-        "rollback_performed": rollback_performed,
-        "install_path": install_path,
-        "error": update_error,
-    });
-
-    if !args.dry_run && update_error.is_some() {
-        RenderDetail::error(summary, json_detail)
-    } else {
-        RenderDetail::with_json(summary, json_detail)
-    }
-}
-
-fn default_manifest_url(channel: &str) -> String {
-    if channel == "nightly" {
-        "https://github.com/VOID-TECHNOLOGY-INC/PyBun/releases/download/nightly/pybun-release.json"
-            .to_string()
-    } else {
-        "https://github.com/VOID-TECHNOLOGY-INC/PyBun/releases/latest/download/pybun-release.json"
-            .to_string()
-    }
-}
-
-fn release_url_for_version(version: &str) -> String {
-    let trimmed = version.trim_start_matches('v');
-    format!(
-        "https://github.com/VOID-TECHNOLOGY-INC/PyBun/releases/tag/v{}",
-        trimmed
-    )
-}
-
-fn python_which(args: &crate::cli::PythonWhichArgs) -> Result<(String, RenderDetail)> {
-    let cache = Cache::new().map_err(|e| eyre!("failed to initialize cache: {}", e))?;
-    let manager = RuntimeManager::new(cache);
-
-    if let Some(version) = &args.version {
-        // Look up a specific version
-        if manager.is_installed(version) {
-            let path = manager.python_binary(version);
-            let json = json!({
-                "version": version,
-                "path": path.display().to_string(),
-                "managed": true,
-            });
-            return Ok((
-                "which".to_string(),
-                RenderDetail::with_json(path.display().to_string(), json),
-            ));
-        }
-
-        // Check if we can find it via env discovery
-        let working_dir = std::env::current_dir()?;
-        if let Ok(env) = find_python_env(&working_dir) {
-            let json = json!({
-                "version": env.version,
-                "path": env.python_path.display().to_string(),
-                "source": format!("{}", env.source),
-                "managed": false,
-            });
-            return Ok((
-                "which".to_string(),
-                RenderDetail::with_json(env.python_path.display().to_string(), json),
-            ));
-        }
-
-        return Err(eyre!(
-            "Python {} is not installed. Use 'pybun python install {}' to install it.",
-            version,
-            version
-        ));
-    }
-
-    // No version specified - show the default Python that would be used
-    let working_dir = std::env::current_dir()?;
-    let env = find_python_env(&working_dir)?;
-
-    let summary = format!("{} (from {})", env.python_path.display(), env.source);
-    let json = json!({
-        "version": env.version,
-        "path": env.python_path.display().to_string(),
-        "source": format!("{}", env.source),
-        "managed": false,
-    });
-
-    Ok(("which".to_string(), RenderDetail::with_json(summary, json)))
-}
-
-// ---------------------------------------------------------------------------
-// pybun init
-// ---------------------------------------------------------------------------
-
-fn sanitize_project_name(name: &str) -> String {
-    let sanitized: String = name
-        .replace([' ', '-'], "_")
-        .chars()
-        .filter(|c| c.is_ascii_alphanumeric() || *c == '_')
-        .collect::<String>()
-        .to_lowercase();
-
-    if sanitized.chars().next().is_some_and(|c| c.is_numeric()) {
-        format!("_{}", sanitized)
-    } else {
-        sanitized
-    }
-}
-
-fn init_project(args: &InitArgs, collector: &mut EventCollector) -> Result<RenderDetail> {
-    let cwd =
-        std::env::current_dir().map_err(|e| eyre!("failed to get current directory: {}", e))?;
-    let pyproject_path = cwd.join("pyproject.toml");
-    let gitignore_path = cwd.join(".gitignore");
-    let readme_path = cwd.join("README.md");
-
-    let default_name = cwd
-        .file_name()
-        .and_then(|n| n.to_str())
-        .map(sanitize_project_name)
-        .filter(|s| !s.is_empty())
-        .unwrap_or_else(|| "my_project".to_string());
-
-    // Resolve arguments (interactive or defaults)
-    let (name, description, python, author, template) = if args.yes {
-        (
-            args.name.clone().unwrap_or(default_name),
-            args.description.clone(),
-            args.python.clone().or_else(|| Some("3.12".to_string())), // Default to 3.12 if not specified
-            args.author.clone(),
-            args.template,
-        )
-    } else {
-        // Interactive mode — requires a terminal
-        if !std::io::stdin().is_terminal() {
-            collector.diagnostic(Diagnostic {
-                level: DiagnosticLevel::Error,
-                code: Some("E_INIT_NOT_INTERACTIVE".to_string()),
-                message: "Interactive prompt requires a terminal".to_string(),
-                file: None,
-                line: None,
-                suggestion: Some(
-                    "Run with --yes to accept defaults non-interactively: pybun init --yes"
-                        .to_string(),
-                ),
-                context: None,
-                exception_type: None,
-                location: None,
-                next_action: None,
-                fix_candidates: None,
-            });
-            return Err(eyre!(
-                "Interactive prompt requires a terminal. Run with --yes to accept defaults non-interactively: pybun init --yes"
-            ));
-        }
-
-        let theme = ColorfulTheme::default();
-
-        // Name
-        let name: String = if let Some(n) = &args.name {
-            n.clone()
-        } else {
-            Input::with_theme(&theme)
-                .with_prompt("Project name")
-                .default(default_name)
-                .interact_text()?
-        };
-
-        // Description
-        let description: Option<String> = if let Some(d) = &args.description {
-            Some(d.clone())
-        } else {
-            let d: String = Input::with_theme(&theme)
-                .with_prompt("Description")
-                .allow_empty(true)
-                .interact_text()?;
-            if d.is_empty() { None } else { Some(d) }
-        };
-
-        // Python Version
-        let python: Option<String> = if let Some(p) = &args.python {
-            Some(p.clone())
-        } else {
-            let p: String = Input::with_theme(&theme)
-                .with_prompt("Python version")
-                .default("3.12".to_string())
-                .interact_text()?;
-            Some(p)
-        };
-
-        // Author
-        let author: Option<String> = if let Some(a) = &args.author {
-            Some(a.clone())
-        } else {
-            let a: String = Input::with_theme(&theme)
-                .with_prompt("Author")
-                .allow_empty(true)
-                .interact_text()?;
-            if a.is_empty() { None } else { Some(a) }
-        };
-
-        // Template
-        let template = args.template; // Can also make this interactive later if needed
-
-        (name, description, python, author, template)
-    };
-
-    // Check main file existence
-    if pyproject_path.exists() {
-        return Err(eyre!(
-            "pyproject.toml already exists at {}",
-            pyproject_path.display()
-        ));
-    }
-
-    // Build pyproject.toml content
-    let mut pyproject = String::new();
-    pyproject.push_str("[project]\n");
-    pyproject.push_str(&format!("name = \"{}\"\n", name));
-    pyproject.push_str("version = \"0.1.0\"\n");
-
-    if let Some(desc) = &description {
-        pyproject.push_str(&format!("description = \"{}\"\n", desc));
-    }
-
-    if let Some(py) = &python {
-        pyproject.push_str(&format!("requires-python = \">={}\"\n", py));
-    }
-
-    if let Some(auth) = &author {
-        pyproject.push_str(&format!("authors = [{{ name = \"{}\" }}]\n", auth));
-    }
-
-    pyproject.push_str("dependencies = []\n");
-    pyproject.push_str("\n[build-system]\n");
-    pyproject.push_str("requires = [\"hatchling\"]\n");
-    pyproject.push_str("build-backend = \"hatchling.build\"\n");
-
-    // Write pyproject.toml
-    fs::write(&pyproject_path, &pyproject)
-        .map_err(|e| eyre!("failed to write pyproject.toml: {}", e))?;
-    let mut files_created = vec![pyproject_path.display().to_string()];
-    let mut files_skipped = vec![];
-
-    // Create .gitignore (with check)
-    if !gitignore_path.exists() {
-        let gitignore_content = r#"# Byte-compiled / optimized / DLL files
-__pycache__/
-*.py[cod]
-*$py.class
-
-# Virtual environments
-.venv/
-venv/
-ENV/
-
-# Distribution / packaging
-dist/
-build/
-*.egg-info/
-
-# PyBun
-pybun.lockb
-.pybun/
-
-# IDE
-.vscode/
-.idea/
-*.swp
-*.swo
-"#;
-        fs::write(&gitignore_path, gitignore_content)
-            .map_err(|e| eyre!("failed to write .gitignore: {}", e))?;
-        files_created.push(gitignore_path.display().to_string());
-    } else {
-        files_skipped.push(gitignore_path.display().to_string());
-    }
-
-    // Create README.md (with check)
-    if !readme_path.exists() {
-        let readme_content = format!("# {}\n\nA Python project.\n", name);
-        fs::write(&readme_path, readme_content)
-            .map_err(|e| eyre!("failed to write README.md: {}", e))?;
-        files_created.push(readme_path.display().to_string());
-    } else {
-        files_skipped.push(readme_path.display().to_string());
-    }
-
-    // Create src layout if package template
-    if matches!(template, InitTemplate::Package) {
-        let package_name = sanitize_project_name(&name);
-        let src_dir = cwd.join("src").join(&package_name);
-        fs::create_dir_all(&src_dir).map_err(|e| eyre!("failed to create src directory: {}", e))?;
-
-        let init_path = src_dir.join("__init__.py");
-        // Safe to overwrite empty init or check? Usually safe to check.
-        if !init_path.exists() {
-            fs::write(&init_path, "").map_err(|e| eyre!("failed to write __init__.py: {}", e))?;
-            files_created.push(init_path.display().to_string());
-        } else {
-            files_skipped.push(init_path.display().to_string());
-        }
-    }
-
-    let summary = format!(
-        "Initialized project '{}' with {} files ({} skipped)",
-        name,
-        files_created.len(),
-        files_skipped.len()
-    );
-
-    Ok(RenderDetail::with_json(
-        summary,
-        json!({
-            "project_name": name,
-            "template": format!("{:?}", template).to_lowercase(),
-            "files_created": files_created,
-            "files_skipped": files_skipped,
-        }),
-    ))
-}
-
-fn run_drift(args: &DriftArgs, collector: &mut EventCollector) -> Result<RenderDetail> {
-    use crate::drift;
-
-    let cwd =
-        std::env::current_dir().map_err(|e| eyre!("failed to get current directory: {}", e))?;
-    let root = if let Some(path) = &args.path {
-        if path.is_absolute() {
-            path.clone()
-        } else {
-            cwd.join(path)
-        }
-    } else {
-        cwd
-    };
-
-    // Require pyproject.toml
-    if !root.join("pyproject.toml").exists() {
-        collector.error_with_code(
-            "E_DRIFT_NO_PYPROJECT",
-            format!("pyproject.toml not found in {}", root.display()),
-            "Run `pybun init` to create a pyproject.toml, or specify a directory with `pybun drift --path <PATH>`.",
-        );
-        return Ok(RenderDetail::error(
-            "pyproject.toml not found".to_string(),
-            json!({
-                "undeclared_imports": [],
-                "unused_declarations": [],
-                "analysis_notes": ["pyproject.toml not found"],
-                "files_scanned": 0
-            }),
-        ));
-    }
-
-    collector.event(EventType::Custom);
-
-    let result = drift::analyze(&root);
-
-    let undeclared_count = result.undeclared_imports.len();
-    let unused_count = result.unused_declarations.len();
-
-    // Surface undeclared imports as warnings
-    for u in &result.undeclared_imports {
-        collector.diagnostic(Diagnostic {
-            level: DiagnosticLevel::Warning,
-            code: Some("W_DRIFT_UNDECLARED_IMPORT".to_string()),
-            message: format!(
-                "Package '{}' is imported but not declared in pyproject.toml",
-                u.package
-            ),
-            file: None,
-            line: None,
-            suggestion: Some(format!("Run `pybun add {}`", u.package)),
-            context: None,
-            exception_type: None,
-            location: None,
-            next_action: None,
-            fix_candidates: None,
-        });
-    }
-
-    // Surface unused declarations as warnings
-    for u in &result.unused_declarations {
-        collector.diagnostic(Diagnostic {
-            level: DiagnosticLevel::Warning,
-            code: Some("W_DRIFT_UNUSED_DECLARATION".to_string()),
-            message: format!(
-                "Package '{}' is declared in pyproject.toml but never imported",
-                u.package
-            ),
-            file: None,
-            line: None,
-            suggestion: Some(format!("Run `pybun remove {}`", u.package)),
-            context: None,
-            exception_type: None,
-            location: None,
-            next_action: None,
-            fix_candidates: None,
-        });
-    }
-
-    let summary = if undeclared_count == 0 && unused_count == 0 {
-        format!("No drift detected ({} files scanned)", result.files_scanned)
-    } else {
-        format!(
-            "Drift detected: {} undeclared import(s), {} unused declaration(s) ({} files scanned)",
-            undeclared_count, unused_count, result.files_scanned
-        )
-    };
-
-    Ok(RenderDetail::with_json(
-        summary,
-        json!({
-            "undeclared_imports": result.undeclared_imports,
-            "unused_declarations": result.unused_declarations,
-            "analysis_notes": result.analysis_notes,
-            "files_scanned": result.files_scanned,
-        }),
-    ))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn parse_package_spec_simple_name() {
-        let (name, version) = parse_package_spec("cowsay");
-        assert_eq!(name, "cowsay");
-        assert_eq!(version, None);
-    }
-
-    #[test]
-    fn parse_package_spec_exact_version() {
-        let (name, version) = parse_package_spec("cowsay==6.1");
-        assert_eq!(name, "cowsay");
-        assert_eq!(version, Some("6.1".to_string()));
-    }
-
-    #[test]
-    fn parse_package_spec_minimum_version() {
-        let (name, version) = parse_package_spec("requests>=2.28.0");
-        assert_eq!(name, "requests");
-        assert_eq!(version, Some("2.28.0".to_string()));
-    }
-
-    #[test]
-    fn parse_package_spec_maximum_version() {
-        let (name, version) = parse_package_spec("numpy<2.0");
-        assert_eq!(name, "numpy");
-        assert_eq!(version, Some("2.0".to_string()));
-    }
-
-    #[test]
-    fn parse_package_spec_compatible_version() {
-        let (name, version) = parse_package_spec("flask~=2.0.0");
-        assert_eq!(name, "flask");
-        assert_eq!(version, Some("2.0.0".to_string()));
-    }
-
-    #[test]
-    fn parse_package_spec_not_equal() {
-        let (name, version) = parse_package_spec("django!=3.0");
-        assert_eq!(name, "django");
-        assert_eq!(version, Some("3.0".to_string()));
     }
 }
