+++ src/commands/tooling.rs
@@ -1,5 +1,6 @@
 use super::RenderDetail;
 use crate::cli::{LazyImportArgs, ModuleFindArgs, ProfileArgs, WatchArgs};
+use crate::env::find_python_env;
 #[cfg(feature = "native-watch")]
 use crate::hot_reload::run_native_watch_loop;
 #[cfg(not(feature = "native-watch"))]
@@ -10,9 +11,14 @@ use crate::lazy_import::{
 };
 use crate::module_finder::{ModuleFinder, ModuleFinderConfig};
 use crate::profiles::{Profile, ProfileConfig, ProfileManager};
+use crate::release_manifest::{ReleaseManifest, current_release_target};
 use crate::schema::EventCollector;
+use crate::self_update::apply_update_for_asset;
 use color_eyre::eyre::{Result, eyre};
 use serde_json::{Value, json};
+use std::cmp::Ordering;
+use std::path::PathBuf;
+use std::process::Command as ProcessCommand;
 
 // ---------------------------------------------------------------------------
 // pybun module-find (Rust-based module finder)
@@ -721,4 +727,460 @@ pub(super) fn run_profile(
             },
         }),
     ))
+}
+
+// ---------------------------------------------------------------------------
+// pybun x (execute tool ad-hoc)
+// ---------------------------------------------------------------------------
+
+#[derive(Debug)]
+pub(super) struct XOutcome {
+    pub(super) summary: String,
+    pub(super) package: String,
+    pub(super) version: Option<String>,
+    pub(super) passthrough: Vec<String>,
+    pub(super) temp_env: String,
+    pub(super) python_version: String,
+    pub(super) exit_code: i32,
+    pub(super) cleanup: bool,
+}
+
+pub(super) fn execute_tool(
+    args: &crate::cli::ToolArgs,
+    _collector: &mut EventCollector,
+) -> Result<XOutcome> {
+    let package_spec = args
+        .package
+        .as_ref()
+        .ok_or_else(|| eyre!("package name is required"))?;
+
+    // Parse package name and version
+    let (package_name, version) = parse_package_spec(package_spec);
+
+    // Check for dry-run mode (for testing)
+    let dry_run = std::env::var("PYBUN_X_DRY_RUN").is_ok();
+
+    // Find Python interpreter
+    let working_dir = std::env::current_dir()?;
+    let env = find_python_env(&working_dir)?;
+    let python_path = env.python_path.to_string_lossy().to_string();
+    let python_version = env.version.clone().unwrap_or_else(|| "unknown".to_string());
+
+    // Create temporary environment
+    let temp_dir =
+        tempfile::tempdir().map_err(|e| eyre!("failed to create temp directory: {}", e))?;
+    let temp_env_path = temp_dir.path().to_string_lossy().to_string();
+
+    if dry_run {
+        // In dry-run mode, just return the planned actions. Tests can set
+        // PYBUN_X_DRY_RUN_EXIT_CODE to simulate a tool that exits non-zero
+        // without needing network access / a real pip install.
+        let exit_code = std::env::var("PYBUN_X_DRY_RUN_EXIT_CODE")
+            .ok()
+            .and_then(|v| v.parse().ok())
+            .unwrap_or(0);
+        return Ok(XOutcome {
+            summary: format!("would execute {} (dry-run)", package_name),
+            package: package_name,
+            version,
+            passthrough: args.passthrough.clone(),
+            temp_env: temp_env_path,
+            python_version,
+            exit_code,
+            cleanup: true,
+        });
+    }
+
+    // Create virtual environment in temp directory
+    let venv_path = temp_dir.path().join("venv");
+    eprintln!(
+        "info: creating temporary environment at {}",
+        venv_path.display()
+    );
+
+    let uv_available = crate::env::find_uv_executable().is_some();
+
+    let mut venv_cmd = ProcessCommand::new(&python_path);
+    venv_cmd.args(["-m", "venv"]);
+    if uv_available {
+        venv_cmd.arg("--without-pip");
+    }
+    venv_cmd.arg(&venv_path);
+    let venv_status = venv_cmd
+        .status()
+        .map_err(|e| eyre!("failed to create virtual environment: {}", e))?;
+
+    if !venv_status.success() {
+        return Err(eyre!("failed to create virtual environment"));
+    }
+
+    // Get pip path in venv
+    let pip_path = if cfg!(windows) {
+        venv_path.join("Scripts").join("pip.exe")
+    } else {
+        venv_path.join("bin").join("pip")
+    };
+
+    // Get python path in venv
+    let venv_python = if cfg!(windows) {
+        venv_path.join("Scripts").join("python.exe")
+    } else {
+        venv_path.join("bin").join("python")
+    };
+
+    // Install the package using uv if available, otherwise pip
+    eprintln!("info: installing {}...", package_spec);
+    let install_status = if let Some(uv_path) = crate::env::find_uv_executable() {
+        ProcessCommand::new(uv_path)
+            .args(["pip", "install", "--quiet", "--python"])
+            .arg(&venv_path)
+            .arg(package_spec)
+            .status()
+            .map_err(|e| eyre!("failed to install package with uv: {}", e))?
+    } else {
+        ProcessCommand::new(&pip_path)
+            .args(["install", "--quiet", package_spec])
+            .status()
+            .map_err(|e| eyre!("failed to install package: {}", e))?
+    };
+
+    if !install_status.success() {
+        return Err(eyre!("failed to install package {}", package_spec));
+    }
+
+    // Find and execute the entry point
+    // Most packages have a console script with the same name as the package
+    let entry_point = if cfg!(windows) {
+        venv_path
+            .join("Scripts")
+            .join(format!("{}.exe", package_name))
+    } else {
+        venv_path.join("bin").join(&package_name)
+    };
+
+    let exit_code = if entry_point.exists() {
+        // Execute the console script directly
+        eprintln!("info: executing {}...", entry_point.display());
+        let mut cmd = ProcessCommand::new(&entry_point);
+        for arg in &args.passthrough {
+            cmd.arg(arg);
+        }
+        let status = cmd
+            .status()
+            .map_err(|e| eyre!("failed to execute {}: {}", package_name, e))?;
+        status.code().unwrap_or(-1)
+    } else {
+        // Fallback: try to run as a module
+        eprintln!("info: executing python -m {}...", package_name);
+        let mut cmd = ProcessCommand::new(&venv_python);
+        cmd.args(["-m", &package_name]);
+        for arg in &args.passthrough {
+            cmd.arg(arg);
+        }
+        let status = cmd
+            .status()
+            .map_err(|e| eyre!("failed to execute module {}: {}", package_name, e))?;
+        status.code().unwrap_or(-1)
+    };
+
+    // Cleanup is automatic when temp_dir is dropped
+    let summary = if exit_code == 0 {
+        format!("executed {} successfully", package_name)
+    } else {
+        format!("{} exited with code {}", package_name, exit_code)
+    };
+
+    Ok(XOutcome {
+        summary,
+        package: package_name,
+        version,
+        passthrough: args.passthrough.clone(),
+        temp_env: temp_env_path,
+        python_version,
+        exit_code,
+        cleanup: true,
+    })
+}
+
+/// Parse a package specification like "cowsay==6.1" into (name, version)
+pub(super) fn parse_package_spec(spec: &str) -> (String, Option<String>) {
+    // Handle various specifier formats
+    for sep in ["==", ">=", "<=", "!=", "~=", ">", "<"] {
+        if let Some(idx) = spec.find(sep) {
+            let name = spec[..idx].to_string();
+            let version = spec[idx + sep.len()..].to_string();
+            return (name, Some(version));
+        }
+    }
+    (spec.to_string(), None)
+}
+
+// ---------------------------------------------------------------------------
+// pybun self update
+// ---------------------------------------------------------------------------
+
+pub(super) fn run_self_update(
+    args: &crate::cli::SelfUpdateArgs,
+    collector: &mut EventCollector,
+) -> RenderDetail {
+    let current_version = env!("CARGO_PKG_VERSION");
+    let channel = &args.channel;
+
+    collector.info(format!("Checking for updates on {} channel", channel));
+
+    let manifest_source_env = std::env::var("PYBUN_SELF_UPDATE_MANIFEST").ok();
+    let default_manifest_url = default_manifest_url(channel);
+    let manifest_source = manifest_source_env
+        .clone()
+        .unwrap_or_else(|| default_manifest_url.clone());
+    let should_fetch_manifest = !args.dry_run
+        || manifest_source_env.is_some()
+        || std::env::var("PYBUN_SELF_UPDATE_FETCH").is_ok();
+    let manifest_result = if should_fetch_manifest {
+        Some(ReleaseManifest::load(&manifest_source))
+    } else {
+        None
+    };
+
+    let mut latest_version = current_version.to_string();
+    let mut update_available = false;
+    let mut release_url = release_url_for_version(current_version);
+    let target = current_release_target();
+    let mut selected_asset = None;
+    let mut manifest_detail = None;
+    let mut manifest_error = None;
+    let mut update_applied = false;
+    let mut rollback_performed = false;
+    let mut install_path = None;
+    let mut update_error = None;
+
+    match manifest_result {
+        Some(Ok(manifest)) => {
+            latest_version = manifest.version.clone();
+            update_available = manifest
+                .compare_version(current_version)
+                .map(|ordering| ordering == Ordering::Greater)
+                .unwrap_or(false);
+            release_url = manifest
+                .release_url
+                .clone()
+                .unwrap_or_else(|| release_url_for_version(&manifest.version));
+
+            selected_asset = target
+                .as_deref()
+                .and_then(|target| manifest.select_asset(target))
+                .cloned();
+            let asset_json = selected_asset
+                .as_ref()
+                .map(|asset| serde_json::to_value(asset).unwrap_or_else(|_| json!({})))
+                .unwrap_or(Value::Null);
+
+            manifest_detail = Some(json!({
+                "version": manifest.version,
+                "channel": manifest.channel,
+                "published_at": manifest.published_at,
+                "release_url": manifest.release_url,
+                "release_notes": manifest.release_notes,
+                "source": manifest_source,
+                "target": target,
+                "asset": asset_json,
+                "assets": manifest.assets.len(),
+                "sbom": manifest.sbom,
+                "provenance": manifest.provenance,
+            }));
+        }
+        Some(Err(error)) => {
+            manifest_error = Some(error.to_string());
+        }
+        None => {}
+    }
+
+    if !args.dry_run {
+        if let Some(error) = manifest_error.as_deref() {
+            let message = format!("failed to load release manifest: {error}");
+            collector.error_with_code(
+                "E_SELF_UPDATE_MANIFEST",
+                message.clone(),
+                "Check network connectivity and the release manifest URL (--channel or PYBUN_SELF_UPDATE_MANIFEST_URL), then retry `pybun self update`.",
+            );
+            update_error = Some(message);
+        } else if update_available {
+            let Some(asset) = selected_asset else {
+                let target_text = target
+                    .as_deref()
+                    .map(|value| value.to_string())
+                    .unwrap_or_else(|| "unknown".to_string());
+                let message = format!("no release asset found for target {target_text}");
+                collector.error_with_code(
+                    "E_SELF_UPDATE_NO_ASSET",
+                    message.clone(),
+                    "Pass --target explicitly to select a supported platform/arch, or check that a release asset exists for your platform.",
+                );
+                update_error = Some(message);
+                let summary = "Update failed: no release asset found".to_string();
+                let json_detail = json!({
+                    "current_version": current_version,
+                    "latest_version": latest_version,
+                    "channel": channel,
+                    "update_available": update_available,
+                    "release_url": release_url,
+                    "dry_run": args.dry_run,
+                    "target": target,
+                    "manifest": manifest_detail,
+                    "manifest_error": manifest_error,
+                    "manifest_source": manifest_source_env.or(Some(default_manifest_url)),
+                    "update_applied": false,
+                    "rollback_performed": false,
+                    "install_path": Value::Null,
+                    "error": update_error,
+                });
+                return RenderDetail::error(summary, json_detail);
+            };
+            let install_override = std::env::var("PYBUN_SELF_UPDATE_BIN")
+                .ok()
+                .map(PathBuf::from);
+            let fail_swap_for_test = std::env::var("PYBUN_SELF_UPDATE_TEST_FAIL_SWAP").is_ok();
+            let target_name = target
+                .as_deref()
+                .unwrap_or(asset.target.as_str())
+                .to_string();
+
+            match apply_update_for_asset(&asset, &target_name, install_override, fail_swap_for_test)
+            {
+                Ok(outcome) => {
+                    update_applied = true;
+                    rollback_performed = outcome.rollback_performed;
+                    install_path = Some(outcome.install_path.display().to_string());
+                    collector.info(format!(
+                        "Updated binary at {}",
+                        outcome.install_path.display()
+                    ));
+                }
+                Err(error) => {
+                    rollback_performed = error.rollback_performed;
+                    update_error = Some(error.to_string());
+                    collector.error_with_code(
+                        "E_SELF_UPDATE_APPLY_FAILED",
+                        error.to_string(),
+                        "Check write permissions to the install path and retry `pybun self update`. If a backup exists, rollback may have already been performed.",
+                    );
+                }
+            }
+        }
+    }
+
+    let summary = if args.dry_run {
+        if update_available {
+            format!(
+                "Update available: {} -> {} (dry-run, no changes made)",
+                current_version, latest_version
+            )
+        } else if let Some(error) = manifest_error.as_deref() {
+            format!("Update check failed: {} (dry-run)", error)
+        } else {
+            format!(
+                "Already up to date: {} (channel: {})",
+                current_version, channel
+            )
+        }
+    } else if let Some(error) = update_error.as_deref() {
+        format!("Update failed: {error}")
+    } else if update_available && update_applied {
+        format!("Updated: {} -> {}", current_version, latest_version)
+    } else if update_available {
+        format!("Update failed: {} -> {}", current_version, latest_version)
+    } else {
+        format!(
+            "Already up to date: {} (channel: {})",
+            current_version, channel
+        )
+    };
+
+    let json_detail = json!({
+        "current_version": current_version,
+        "latest_version": latest_version,
+        "channel": channel,
+        "update_available": update_available,
+        "release_url": release_url,
+        "dry_run": args.dry_run,
+        "target": target,
+        "manifest": manifest_detail,
+        "manifest_error": manifest_error,
+        "manifest_source": manifest_source_env.or(Some(default_manifest_url)),
+        "update_applied": update_applied,
+        "rollback_performed": rollback_performed,
+        "install_path": install_path,
+        "error": update_error,
+    });
+
+    if !args.dry_run && update_error.is_some() {
+        RenderDetail::error(summary, json_detail)
+    } else {
+        RenderDetail::with_json(summary, json_detail)
+    }
+}
+
+fn default_manifest_url(channel: &str) -> String {
+    if channel == "nightly" {
+        "https://github.com/VOID-TECHNOLOGY-INC/PyBun/releases/download/nightly/pybun-release.json"
+            .to_string()
+    } else {
+        "https://github.com/VOID-TECHNOLOGY-INC/PyBun/releases/latest/download/pybun-release.json"
+            .to_string()
+    }
+}
+
+fn release_url_for_version(version: &str) -> String {
+    let trimmed = version.trim_start_matches('v');
+    format!(
+        "https://github.com/VOID-TECHNOLOGY-INC/PyBun/releases/tag/v{}",
+        trimmed
+    )
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn parse_package_spec_simple_name() {
+        let (name, version) = parse_package_spec("cowsay");
+        assert_eq!(name, "cowsay");
+        assert_eq!(version, None);
+    }
+
+    #[test]
+    fn parse_package_spec_exact_version() {
+        let (name, version) = parse_package_spec("cowsay==6.1");
+        assert_eq!(name, "cowsay");
+        assert_eq!(version, Some("6.1".to_string()));
+    }
+
+    #[test]
+    fn parse_package_spec_minimum_version() {
+        let (name, version) = parse_package_spec("requests>=2.28.0");
+        assert_eq!(name, "requests");
+        assert_eq!(version, Some("2.28.0".to_string()));
+    }
+
+    #[test]
+    fn parse_package_spec_maximum_version() {
+        let (name, version) = parse_package_spec("numpy<2.0");
+        assert_eq!(name, "numpy");
+        assert_eq!(version, Some("2.0".to_string()));
+    }
+
+    #[test]
+    fn parse_package_spec_compatible_version() {
+        let (name, version) = parse_package_spec("flask~=2.0.0");
+        assert_eq!(name, "flask");
+        assert_eq!(version, Some("2.0.0".to_string()));
+    }
+
+    #[test]
+    fn parse_package_spec_not_equal() {
+        let (name, version) = parse_package_spec("django!=3.0");
+        assert_eq!(name, "django");
+        assert_eq!(version, Some("3.0".to_string()));
+    }
 }
